% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_arqtra.R
\name{import_arqtra}
\alias{import_arqtra}
\title{Importar arquivo das transações monetárias de cartões Visa}
\usage{
import_arqtra(
  caminho,
  import_db = "none",
  host = "32.30.10.224",
  port = 5432,
  usuario,
  senha,
  db,
  tabela = "arqtra",
  metodo = "vroom"
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{import_db}{String. "none", "duckdb" ou "postgres".}

\item{host}{String. Se import_db = "postgres", informar o IP do servidor.}

\item{port}{Integer. Se import_db = "postgres", informar a porta do servidor.}

\item{usuario}{String. Se import_db = "postgres", o usuário deverá ser informado.}

\item{senha}{String. Se import_db = "postgres", a senha deverá ser informada.}

\item{db}{String. Se duckdb = TRUE, o caminho do arquivo do banco de dados.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}

\item{metodo}{String. "readr" ou "vroom".}
}
\description{
Importa o arquivo ARQTRA para o R, para o DuckDB ou para o Postgres.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.

No caso de importar para o Postgres será necessário um usuário e uma senha e
o servidor deverá estar escutando o IP do usuário.
}
\examples{
\dontrun{
# para importar para o R:
data = import_arqtra("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_arqtra(
 "\\\\path\\\\to\\\\file.txt",
 import_db = "duckdb",
 db = "\\\\path\\\\to\\\\duck.db"),
 tabela = "arqtra"
}

# para importar para o Postgres:
import_arqtra(
"\\\\path\\\\to\\\\file.txt"),
import_db = "postgres",
usuario = "user",
senha = "password",
tabela = "arqtra")

}
