% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_banescard.R
\name{import_banescard}
\alias{import_banescard}
\title{Importar arquivo de transações do cartão Banescard}
\usage{
import_banescard(
  caminho,
  import_db = "none",
  host = "32.30.10.224",
  port = 5432,
  usuario,
  senha,
  db,
  tabela = "banescard",
  metodo = "vroom"
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.csv).}

\item{import_db}{String. "none", "duckdb" ou "postgres".}

\item{host}{String. Se import_db = "postgres", informar o IP do banco de dados.}

\item{port}{Int. Se import_db = "postgres", informar a porta.}

\item{usuario}{String. Se import_db = "postgres", o usuário deverá ser informado.}

\item{senha}{String. Se import_db = "postgres", a senha deverá ser informada.}

\item{db}{String. Se import_db = "duckdb", o caminho do arquivo do banco de dados.
Se import_db = "postgres", o nome do DB.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}

\item{metodo}{String. "readr" ou "vroom".}
}
\description{
Importar o arquivo Banescard para o R, para o DuckDB ou para o Postgres.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.

No caso de importar par ao Postgres será necessário um usuário e uma senha.
Além disso, o IP também deve ser autorizado.
}
\examples{
\dontrun{
# para importar para o R:
data = import_banescard("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_banescard(
"\\\\path\\\\to\\\\file.txt",
import_db = "duckdb",
db = "\\\\path\\\\to\\\\duck.db",
tabela = "banescard")

# para importar para o Postgres:
import_banescard(
"\\\\path\\\\to\\\\file.txt",
import_db = "postgres",
usuario = "user",
senha = "password",
tabela = "banescard")
}

}
