% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dco.R
\name{import_dco}
\alias{import_dco}
\title{Importar arquivo da base geral de clientes do DCO.}
\usage{
import_dco(
  caminho,
  import_db = "none",
  host = "32.30.10.224",
  port = 5432,
  usuario,
  senha,
  metodo = "vroom",
  db,
  tabela
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{import_db}{String. "none", "duckdb" ou "postgres".}

\item{host}{String. Se import_db = "postgres", informar o IP do banco de dados.}

\item{port}{Int. Se import_db = "postgres", informar a porta.}

\item{usuario}{String. Se import_db = "postgres", o usuário deverá ser informado.}

\item{senha}{String. Se import_db = "postgres", a senha deverá ser informada.}

\item{metodo}{String. "readr" ou "vroom".}

\item{db}{String. Se import_db = "duckdb", o caminho do arquivo do banco de dados.
Se import_db = "postgres", o nome do DB.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}
}
\description{
Importa o arquivo DCO para o R, para o DuckDB ou para o Postgres.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.

No caso de importar par ao Postgres será necessário um usuário e uma senha.
Além disso, o IP também deve ser autorizado.
}
\examples{
\dontrun{
# para importar para o R:
data <- import_dco("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_dco(
  "\\\\path\\\\to\\\\file.txt",
  import_db = "duckdb",
  db = "\\\\path\\\\to\\\\duck.db",
  tabela = dpvs_202110
)

# para importar para o Postgres:
import_dco(
  "\\\\path\\\\to\\\\file.txt",
  import_db = "postgres",
  usuario = "user",
  senha = "password",
  db = "banestes",
  tabela = "dpvs_202110"
)
}

}
