% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pix.R
\name{import_pix}
\alias{import_pix}
\title{Importar arquivo das transações de PIX}
\usage{
import_pix(
  caminho,
  tipo,
  host = "32.30.10.224",
  port = 5432,
  import_db = "none",
  usuario,
  senha,
  db,
  tabela,
  metodo = "vroom"
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{host}{String. Se import_db = "postgres", informar o IP do servidor.}

\item{port}{Integer. Se import_db = "postgres", informar a porta do servidor.}

\item{import_db}{String. "none", "duckdb" ou "postgres".}

\item{usuario}{String. Se import_db = "postgres", o usuário deverá ser informado.}

\item{senha}{String. Se import_db = "postgres", a senha deverá ser informada.}

\item{db}{String. Se import_db = "duckdb", o caminho do arquivo do banco de dados.
Se import_db = "postgres", o nome do DB.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}

\item{metodo}{String. "readr" ou "vroom".}
}
\description{
Importa o arquivo PIX para o R, para o DuckDB ou para o Postgres.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.

No caso de importar para o Postgres será necessário um usuário e uma senha e
o servidor deverá estar escutando o IP do usuário.
}
\examples{
\dontrun{
# para importar para o R:
data = import_pix("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_pix(
 "\\\\path\\\\to\\\\file.txt",
 import_db = "duckdb",
 db = "\\\\path\\\\to\\\\duck.db",
 tabela = "pix")

import_pix(
"\\\\path\\\\to\\\\file.txt",
import_db = "postgres",
usuario = "user",
senha = "passoword",
db = "banestes",
tabela = c("pix_recebimento", "pix_pagamento"))
}

}
