% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scr.R
\name{import_scr}
\alias{import_scr}
\title{Importar arquivo da base geral do Banco Central}
\usage{
import_scr(
  caminho,
  import_db = "none",
  host = "32.30.10.224",
  port = 5432,
  usuario,
  senha,
  db,
  tabela
)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.csv).}

\item{import_db}{String. "none", "duckdb" ou "postgres".}

\item{host}{String. Se import_db = "postgres", informar o IP do banco de dados.}

\item{port}{Int. Se import_db = "postgres", informar a porta.}

\item{usuario}{String. Se import_db = "postgres", o usuário deverá ser informado.}

\item{senha}{String. Se import_db = "postgres", a senha deverá ser informada.}

\item{db}{String. Se duckdb = TRUE, o caminho do arquivo do banco de dados.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}
}
\description{
Importa o arquivo scr para o R, para o DuckDB ou para o Postgres.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.

No caso de importar par ao Postgres será necessário um usuário e uma senha.
Além disso, o IP também deve ser autorizado.
}
\examples{
\dontrun{
# para importar para o R:
data = import_arqtra("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_scr(
"\\\\path\\\\to\\\\file.scv",
import_db = "duckdb",
db = "\\\\path\\\\to\\\\duck.db"),
tabela = "scr"
}

# para importar para o Postgres:
import_scr(
"\\\\path\\\\to\\\\file.txt"),
import_scr = "postgres",
usuario = "user",
senha = "password",
tabela = "scr")

}
