% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcamento_ag.R
\name{orcamento_ag}
\alias{orcamento_ag}
\title{Distribuir orcamento para agencias}
\usage{
orcamento_ag(caminho, tipo, previsao)
}
\arguments{
\item{caminho}{String. O caminho do orcamento realizado do ano anterior,
incluindo a extensao (.txt).}

\item{tipo}{String. "resultado" ou "saldo".}

\item{previsao}{Numeric. Vetor das previsoes para o ano seguinte.}
}
\description{
Distribui proporcionalmente o valor orcado para a rede de agencias
em funcao do total do ano ou do ultimo mes.
}
\details{
O vetor de previsoes deve ter 12 valores, um para cada mes.
O arquivo a ser importado deve ser baixado do PPO, aberto e convertido
para xlsx.
}
\examples{
\dontrun{
# vetor de previsao
previsao = rep(1000, 12)
# obter dataframe
orcamento_ag(
caminho = "\\\\path\\\\to\\\\file.xlsx"),
tipo = "saldo",
previsao = previsao
}

}
