% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dco.R
\name{import_dco}
\alias{import_dco}
\title{Importar arquivo da base geral de clientes do DCO}
\usage{
import_dco(caminho, metodo = "vroom", duckdb = FALSE, db, tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{metodo}{String. "readr" ou "vroom".}

\item{duckdb}{Logical.
Se TRUE, entao o arquivo sera importado para o DuckDB.}

\item{db}{String. Se duckdb = TRUE, o caminho do arquivo do banco de dados.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}
}
\description{
Importa o arquivo DCO para o R ou para o DuckDB.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_dco("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_dco(
 "\\\\path\\\\to\\\\file.txt",
 duckdb = TRUE,
 db = "\\\\path\\\\to\\\\duck.db")
}

}
