% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_x4.R
\name{import_x4}
\alias{import_x4}
\title{Importar arquivo de folha de pagamento}
\usage{
import_x4(caminho, metodo = "vroom", duckdb = FALSE, db, tabela)
}
\arguments{
\item{caminho}{String. O caminho do arquivo a ser importado, incluindo a
extensao (.txt).}

\item{metodo}{String. "readr" ou "vroom".}

\item{duckdb}{Logical.
Se TRUE, entao o arquivo sera importado para o DuckDB.}

\item{db}{String. Se duckdb = TRUE, o caminho do arquivo do banco de dados.}

\item{tabela}{String. Nome da tabela a ser criada no DuckDB.}
}
\description{
Importa o arquivo X4 do sistema DPI para o R ou para o DuckDB.
}
\details{
Se for importar apenas para o R, lembre-se de nomear um objeto para
o resultado da funcao. Caso for importar para o DuckDB, nao e necessario
criar um objeto. Note que a tabela sera sobrescrita se ja existir.
}
\examples{
\dontrun{
# para importar para o R:
data = import_x4("\\\\path\\\\to\\\\file.txt")

# para importar para o DuckDB:
import_x4(
 "\\\\path\\\\to\\\\file.txt",
 duckdb = TRUE,
 db = "\\\\path\\\\to\\\\duck.db"),
 tabela = "t_201901"
}

}
